[ONNX] Trace model if quantization is detected
Previously pre-tracing model is required for exporting quantized model.
e.g. calling `traced_m = torch.jit.trace(model, inputs)` and export `traced_m`.
The reason was quantized weights are stored in a unique `PackedParam` structure,
and they need to be handled by tracing to be exportable.
This PR enables export api to call tracing underneath if it detects quantization
in the model.

ghstack-source-id: b3f542bab12208166bc9450b055d1056cfec9937
Pull Request resolved: https://github.com/pytorch/pytorch/pull/75921